src/functions/csv-to-bigquery/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
